feat #3: Theater class

+++ src/data/models/spm/activityTypes/Theater.cpp
@@ -1,14 +1,14 @@
 #include "Theater.h"
 
-Theater::Theater(const string &name, const string &description, int duration, double price) : Activity(name,
+Theater::Theater(const string &name, const string &description, int duration, float price) : Activity(name,
                                                                                                        description,
                                                                                                        duration,
                                                                                                        price) {}
 
-Theater::Theater(const string &name, const string &description, int duration, double price, int capacity) : Activity(
+Theater::Theater(const string &name, const string &description, int duration, float price, int capacity) : Activity(
         name, description, duration, price, capacity) {}
 
-double Theater::getPrice(int age) const {
+float Theater::getPrice(int age) const {
     if (age <= MAX_YOUTH_THEATER_DISCOUNT_AGE) {
         return Activity::getPrice() - Activity::getPrice() * YOUTH_THEATER_DISCOUNT;
     } else if (age >= MIN_SENIOR_CITIZEN_THEATER_DISCOUNT_AGE) {
